+++ packages/numbers/index.js
@@ -1,11 +1,11 @@
 // import packages
 const isEven = require('../../utils/numbers/isEven')
-const round = require('../../utils/numbers/round')
+const roundTo = require('../../utils/numbers/roundTo')
 const toReadable = require('../../utils/numbers/toReadable')
 
 // export packages
 module.exports = {
 	isEven,
-	round,
+	roundTo,
 	toReadable,
 }
